feat: implement list in table (fix #321) (#491)
* feat: implement to maintain tag inserted in markdown

* feat: implement to insert list in table

* refactor: reflect code review

* fix: key(up/down, enter, backspace, delete) operation in list of table (#492)

* fix: up and down key does not work in list of table

* fix: do not block squire to merge td

When press backspace and delete in end of the td or empty td,
Squire merge TD or TR. So we block backspace and delete in these case.
But if TD or TR except when Squire merge block,
those code (blocking backspace and delete key event) is not needed.
So remove those code and fix Squire by handle TD and TR to exception
case.

* fix: can not remove list item when backspace occur first list item in table

* fix: div is generated when enter occur in empty li

* fix: delete key in the end of the list

* refactor: reflect code review

* chore: squire update (squire do not merge table)

* refactor: reflect code review